Cleanup property keys and references

+++ packages/core/src/evaluator/node-evaluators/UnaryExpression.ts
@@ -5,11 +5,13 @@ import { getValue } from "../../runtime/algorithms/get-value.js";
 import { toBoolean } from "../../runtime/algorithms/to-boolean.js";
 import { toNumber } from "../../runtime/algorithms/to-number.js";
 import { toObject } from "../../runtime/algorithms/to-object.js";
+import { toPropertyKey } from "../../runtime/algorithms/to-property-key.js";
+import { isPrivateReference } from "../../runtime/references/is-private-reference.js";
 import { isPropertyReference } from "../../runtime/references/is-property-reference.js";
+import { isSuperReference } from "../../runtime/references/is-super-reference.js";
 import { isUnresolvableReference } from "../../runtime/references/is-unresolvable-reference.js";
 import { isStaticJsReferenceRecord } from "../../runtime/references/StaticJsReferenceRecord.js";
 import { isStaticJsValue } from "../../runtime/types/StaticJsValue.js";
-import { toPropertyKey } from "../../runtime/utils/to-property-key.js";
 import { EvaluateNodeCommand } from "../commands/EvaluateNodeCommand.js";
 import { Completion } from "../completions/Completion.js";
 import { Q } from "../completions/Q.js";
@@ -70,7 +72,14 @@ function* deleteExpressionNodeEvaluator(node: UnaryExpression): EvaluationGenera
   }
 
   if (isPropertyReference(ref)) {
-    // TODO: if is super reference, throw.
+    if (isPrivateReference(ref)) {
+      throw new StaticJsEngineError("Cannot delete a private reference.");
+    }
+
+    if (isSuperReference(ref)) {
+      throw Completion.Throw("ReferenceError", "Cannot delete a super reference.");
+    }
+
     const baseObj = yield* toObject(ref.base);
     const propertyKey = yield* toPropertyKey(ref.referencedName);
     const result = yield* baseObj.deleteEvaluator(propertyKey);
